+++ website/siteConfig.js
@@ -53,6 +53,10 @@ module.exports = {
       path: '/examples/jsx-column',
     },
     {
+      title: 'Tooltip Cell',
+      path: '/examples/tooltip-cell',
+    },
+    {
       title: 'Auto Resize',
       path: '/examples/auto-resize',
     },
@@ -87,6 +91,10 @@ module.exports = {
     {
       title: 'Hide Header',
       path: '/examples/hide-header',
+    },
+    {
+      title: 'Draggable Rows',
+      path: '/examples/draggable-rows',
     },
     {
       title: 'Multi Header',
@@ -161,6 +169,10 @@ module.exports = {
       path: '/examples/multi-sort',
     },
     {
+      title: 'Selection',
+      path: '/examples/selection',
+    },
+    {
       title: 'Expand - Uncontrolled',
       path: '/examples/expand-uncontrolled',
     },
@@ -187,6 +199,10 @@ module.exports = {
     {
       title: 'Tag Name',
       path: '/examples/tag-name',
+    },
+    {
+      title: 'Inline Editing',
+      path: '/examples/inline-editing',
     },
   ],
 }