+++ website/src/components/Header.js
@@ -15,9 +15,9 @@ const Container = styled.div`
 
 const Nav = styled.div`
   margin: 0 auto;
-  max-width: 1000px;
-  height: 50px;
-  padding: 0 20px;
+  max-width: 100rem;
+  height: 5rem;
+  padding: 0 2rem;
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -25,8 +25,9 @@ const Nav = styled.div`
 
 const Title = styled(Link)`
   text-decoration: none;
-  font-size: 24px;
-  padding: 10px 0;
+  font-size: 2.4rem;
+  line-height: 1;
+  padding: 1rem 0;
   &,
   &:hover,
   &:focus {
@@ -34,14 +35,17 @@ const Title = styled(Link)`
   }
 `
 
+const Spacer = styled.div`
+  flex: 1;
+`
+
 const NavLink = styled(Link).attrs({
   partiallyActive: true,
 })`
   color: #bcc9d1;
   text-decoration: none;
-  font-size: 16px;
-  line-height: 20px;
-  padding: 10px;
+  padding: 1rem;
+  line-height: 1;
   &:hover {
     color: #fff;
   }
@@ -59,50 +63,43 @@ const NavLink = styled(Link).attrs({
 const ExternalLink = NavLink.withComponent('a')
 
 const LinkIcon = styled.img`
-  width: 20px;
-  height: 20px;
+  width: 2rem;
+  height: 2rem;
 `
 
 const Version = styled(ExternalLink)`
-  font-size: 14px;
+  font-size: 1.4rem;
+  padding: 1rem;
+  margin-top: 1rem;
 `
 
 const Header = ({ pathname }) => {
   return (
     <Container>
       <Nav>
-        <div>
-          <Title to="/">BaseTable</Title>
-          <Version
-            href="https://github.com/Autodesk/react-base-table/releases"
-            target="_blank"
-          >
-            v{pkg.version}
-          </Version>
-        </div>
-        <div>
-          <NavLink to="/docs" pathname={pathname}>
-            Docs
-          </NavLink>
-          <NavLink to="/api" pathname={pathname}>
-            API
-          </NavLink>
-          <NavLink to="/examples" pathname={pathname}>
-            Examples
-          </NavLink>
-          <NavLink to="/playground" pathname={pathname}>
-            Playground
-          </NavLink>
-          <ExternalLink
-            href="https://github.com/Autodesk/react-base-table/blob/master/CHANGELOG.md"
-            target="_blank"
-          >
-            CHANGELOG
-          </ExternalLink>
-          <ExternalLink last href={pkg.repository.url} target="_blank">
-            <LinkIcon src={linkIcon} alt="Github" />
-          </ExternalLink>
-        </div>
+        <Title to="/">BaseTable</Title>
+        <Version
+          href="https://github.com/Autodesk/react-base-table/blob/master/CHANGELOG.md"
+          target="_blank"
+        >
+          v{pkg.version}
+        </Version>
+        <Spacer />
+        <NavLink to="/docs" pathname={pathname}>
+          Docs
+        </NavLink>
+        <NavLink to="/api" pathname={pathname}>
+          API
+        </NavLink>
+        <NavLink to="/examples" pathname={pathname}>
+          Examples
+        </NavLink>
+        <NavLink to="/playground" pathname={pathname}>
+          Playground
+        </NavLink>
+        <ExternalLink last href={pkg.repository.url} target="_blank">
+          <LinkIcon src={linkIcon} alt="Github" />
+        </ExternalLink>
       </Nav>
     </Container>
   )