+++ website/src/components/Methods.js
@@ -5,31 +5,24 @@ import Html from './Html'
 import Anchor from './Anchor'
 
 const Method = styled.div`
-  margin-bottom: 15px;
+  margin-bottom: 1.6rem;
 `
 
 const Name = styled.div`
-  font-size: 14px;
   font-weight: 600;
-  margin-bottom: 5px;
 `
 
 const Tag = styled.span`
-  font-size: 12px;
-  padding: 2px 6px;
-  margin: 0 8px;
-  border-radius: 4px;
+  font-size: 0.8em;
+  padding: 0.1em 0.4em;
+  margin: 0 0.4em;
+  border-radius: 0.2em;
   background-color: #daf0f9;
   color: #819099;
 `
 
 const Block = styled(Html)`
-  font-size: 14px;
-
-  p {
-    font-size: 14px;
-    margin: 0 0 10px;
-  }
+  color: #666;
 `
 
 const getSignature = params =>
@@ -42,7 +35,7 @@ const Methods = ({ title = 'Methods', methods, ...rest }) => (
     <Anchor>{title}</Anchor>
     {Array.isArray(methods) &&
       methods.map((method, idx) => (
-        <Method key={method.name}>
+        <Method key={method.name} id={`methods-${method.name}`}>
           <Name>
             {method.name}
             <Tag>{getSignature(method.params)}</Tag>