+++ website/src/templates/api.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import styled from 'styled-components'
 import { graphql } from 'gatsby'
 
 import Page from 'components/Page'
 import Html from 'components/Html'
 import Props from 'components/Props'
 import Methods from 'components/Methods'
+import Pagination from 'components/Pagination'
 
 import siteConfig from 'siteConfig'
 
@@ -15,24 +15,13 @@ const links = siteConfig.api.map(item => ({
   to: item.path,
 }))
 
-const Title = styled.div`
-  font-size: 16px;
-  font-weight: 700;
-  margin-top: 16px;
-`
-
-const Block = styled(Html)`
-  font-size: 14px;
-  p {
-    font-size: 14px;
-  }
-`
-
 class ApiTemplate extends React.Component {
   render() {
-    const { data, location } = this.props
+    const { data, pageContext, location } = this.props
     const metaData = data.componentMetadata
     const methods = metaData.childrenComponentMethodExt
+    const name = pageContext.name
+    const link = links.find(link => link.to === `/api/${name.toLowerCase()}`)
 
     return (
       <Page
@@ -40,12 +29,13 @@ class ApiTemplate extends React.Component {
         location={location}
         links={links}
       >
-        <Title>{metaData.displayName}</Title>
+        <h1>{metaData.displayName}</h1>
         {metaData.description && (
-          <Block htmlAst={metaData.description.childMarkdownRemark.htmlAst} />
+          <Html htmlAst={metaData.description.childMarkdownRemark.htmlAst} />
         )}
         <Props props={metaData.props} />
         {methods.length > 0 && <Methods methods={methods} />}
+        <Pagination links={links} link={link} />
       </Page>
     )
   }