+++ website/src/templates/example.js
@@ -1,10 +1,9 @@
 import React from 'react'
-import styled from 'styled-components'
 import { graphql } from 'gatsby'
 
-import Anchor from 'components/Anchor'
 import Page from 'components/Page'
 import CodePreview from 'components/CodePreview'
+import Pagination from 'components/Pagination'
 
 import siteConfig from 'siteConfig'
 
@@ -14,12 +13,6 @@ const links = siteConfig.examples.map(item => ({
   to: item.path,
 }))
 
-const Title = styled.div`
-  font-size: 16px;
-  font-weight: 700;
-  margin-top: 16px;
-`
-
 class ComponentTemplate extends React.Component {
   render() {
     const { data, pageContext, location } = this.props
@@ -28,9 +21,9 @@ class ComponentTemplate extends React.Component {
     const link = links.find(link => link.to === `/examples/${name}`)
     return (
       <Page title={`Examples: ${link.title}`} location={location} links={links}>
-        <Title>{link.title}</Title>
-        <Anchor title="Example" />
+        <h1>{link.title}</h1>
         <CodePreview code={code} />
+        <Pagination links={links} link={link} />
       </Page>
     )
   }